chore: restore codeql fixes

+++ src/common/physics/engine.c
@@ -196,7 +196,7 @@ CmdFX_Sprite** Engine_tick() {
 
     int ground = Engine_getGroundY();
     int width = Canvas_getWidth();
-    int forceOfGravity = Engine_getForceOfGravity();
+    double forceOfGravity = Engine_getForceOfGravity();
 
     CmdFX_Sprite** modified =
         calloc(count - _staticSpriteCount + 1, sizeof(CmdFX_Sprite*));
